Move interpolation of next-period policy and value to top-level function: solve_dcegm(), so that next-period policy and value don't need to be accsessed in EGM step.
